drivers: mspi: mspi_dw: Add display support for nrf-mspi
The nrf-mspi has display registers in the wrapper that
need to be configured. Since the bits-per-pixel is set
during the tx configuration, a varible was added to the
nrf-mspi binding. It is expected that other
implementations that support QSPI or DSPI display will
also need to tell their wrapper the bits-per-pixel so the
IP knows how to structure the pixel data. A check was
added to the vendor specific code for these values since
this is what our nrf-mspi IP supports.

Signed-off-by: David Jewsbury <david.jewsbury@nordicsemi.no>